1-list_len.c file

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -9,7 +9,7 @@
 
 size_t list_len(const list_t *h)
 {
-	int element = 0
+	size_t element = 0
 
 		while (h != NULL)
 		{
